requirements.txt
flask
flask-cors
bcrypt
psycopg2
python-dotenv
